* FINALLY, a fix for client side buff timers getting out of step. This turned out to be caused by the 'heartbeat' message, which as part of it's payload sends the current time **elapsed since the server started**. However, the client uses this for its lag prediction, and it actually expects the time **elapsed since the last heartbeat**. This might also explain a piece code code that stops heartbeats being sent while channeling with a comment by Greth 'Receiving a heartbeat message during casts seems to interrupt animations.'. This is likely the cause as the client wouldn't know the correct time. The bug would show times of buffs increasing at an exponential rate every 30 seconds.
 * Redesigned loging page. The news section is smaller (and now supports links). The 'create account' section has been removed and replaced with a single button. And the login section has been shrunk vertically. This all allows the logo AND news to be displayed on the login.

* Added Speed tomes (and another rework of such XP tomes). XP tomes, Speed tomes (and potentially) other items now correctly use the ability system alongside the limited lifetime items (this means world buff icons always work on the client). mActivationAbilityId is now used, as signals an item should be activated as soon it arrives in the inventory (as the buff removed when the item is removed from inventory). Grinders continue to work in the same way as they do not require a buff.

+++ SOURCE/Server/Util.h
@@ -57,7 +57,7 @@ int PrepExt_QueryResponseStringList(char *buffer, int queryIndex, const STRINGLI
 int PrepExt_QueryResponseMultiString(char *buffer, int queryIndex, const MULTISTRING &strData);
 int PrepExt_QueryResponseError(char *buffer, int queryIndex, const char *message);
 int PrepExt_SendEffect(char *buffer, int sourceID, const char *effectName, int targetID);
-int PrepExt_SendHeartbeatMessage(char *buffer, unsigned int elapsedMilliseconds);
+int PrepExt_SendHeartbeatMessage(char *buffer, unsigned long elapsedMilliseconds);
 int PrepExt_SendAdvancedEmote(char *buffer, int creatureID, const char *emoteName, float emoteSpeed, int loop);
 int PrepExt_SendEmoteControl(char *buffer, int creatureID, int emoteEvent);
 int PrepExt_ModStopSwimFlag(char *buffer);
